+++ main.go
@@ -2,12 +2,15 @@ package main
 
 import (
 	"github.com/labstack/echo"
-	"github.com/souko2525/KanCoreAPI/database"
-	"html/template"
-	"io"
-	"net/http"
+	"github.com/labstack/echo/middleware"
+	"github.com/souko2525/KanCoreAPI/router"
+	//"github.com/souko2525/KanCoreAPI/database"
+	//"html/template"
+	//"io"
+	//"net/http"
 )
 
+/*
 type Template struct {
 	templates *template.Template
 }
@@ -19,16 +22,26 @@ func (t *Template) Render(w io.Writer, name string, data interface{}, c echo.Con
 func Hello(c echo.Context) error {
 	return c.Render(http.StatusOK, "layout.html", "echo server")
 }
+*/
 
 func main() {
-	t := &Template{
-		templates: template.Must(template.ParseFiles("templates/layout.html", "templates/hello.html")),
-	}
+	/*
+		t := &Template{
+			templates: template.Must(template.ParseFiles("templates/layout.html", "templates/hello.html")),
+		}
+	*/
 	e := echo.New()
-	e.Renderer = t
-	e.GET("/", func(c echo.Context) error {
-		return c.String(http.StatusOK, "index page")
-	})
-	e.GET("/hello", Hello)
-	e.Logger.Fatal(e.Start(":1323"))
+	e.Use(middleware.Recover())
+	e.Use(middleware.CORS())
+	router.Router(e)
+	e.Start(":3000")
+
+	/*
+		e.Renderer = t
+		e.GET("/", func(c echo.Context) error {
+			return c.String(http.StatusOK, "index page")
+		})
+		e.GET("/hello", Hello)
+		e.Logger.Fatal(e.Start(":1323"))
+	*/
 }
